Moved rest of utils to utils directory
Moved middleware.ts, customErrors.ts and types.ts to server/src/utils
