+++ server/src/utils/customErrors.ts
