+++ server/src/utils/types.ts
@@ -1,5 +1,5 @@
 import {type Request} from 'express';
-import {type UserPublic} from './validators/UserPublic';
+import {type UserPublic} from '../validators/UserPublic';
 
 export type RequestWithUser = {
 	user?: UserPublic;
